增加牌桌ID，牌桌名称，筹码兑换，归还筹码的参数对象

+++ src/main/java/com/dx/facade/account/req/WalletBalanceDTO.java
@@ -37,7 +37,7 @@ public class WalletBalanceDTO implements BaseParmDTO {
     private Integer userType;
 
     @ApiModelProperty("牌桌ID")
-    private Integer tableId;
+    private Long tableId;
 
     @ApiModelProperty("牌桌名称")
     private String tableName;
